MI-405: API endpoint for grt survey
